requirements.txt
labjack-ljm
fastavro